Add pictures for all of us.
Candan, Massimo and Nicola didn't send any other picture (and their picture
also matches the one used in our presentation during the BMS orientation),
so for them it was a natural choice.
In order to look uniform, they should all have the same aspect ratio, so that
Michael and Simona's pictures would have had to be cropped---and the BMS
homepage has them cropped decently already.

If desired, they can still be changed later.